Split input validation test into files
